requirements.txt
Logbook==1.1.0
lxml==4.1.1
py==1.4.34
pyshark==0.3.7.11
trollius==1.0.4
python-libnmap==0.7.0
